Project/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
anyascii==0.3.0; python_version >= '3.3'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0
asgiref==3.4.1; python_version >= '3.6'
async-timeout==4.0.1; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9; python_version >= '3.5'
click==8.0.3; python_version >= '3.6'
coderedcms==0.22.3
colorama==0.4.4; platform_system == 'Windows'
cryptography==36.0.1
decorator==5.1.0; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-admin-honeypot==1.1.0
django-allauth==0.47.0
django-bootstrap4==2.3.1; python_version >= '3.6' and python_version < '4.0'
django-environ==0.8.1
django-eventtools==1.0.3
django-extensions==3.1.5
django-filter==2.4.0; python_version >= '3.5'
django-honeypot==1.0.2
django-modelcluster==5.2; python_version >= '3.5'
django-password-validators==1.6.1
django-ratelimit==3.0.1
django-restricted-sessions==0.3.0
django-taggit==1.5.1; python_version >= '3.6'
django-treebeard==4.5.1; python_version >= '3.6'
django-user-sessions==1.7.1
django-xforwardedfor-middleware==2.0
django[argon2]==3.2.10
djangorestframework==3.13.1; python_version >= '3.6'
draftjs-exporter==2.1.7
et-xmlfile==1.1.0; python_version >= '3.6'
frozenlist==1.2.0; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
geocoder==1.38.1
geoip2==4.5.0
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
icalendar==4.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3.0'
l18n==2021.3
maxminddb==2.2.0; python_version >= '3.6'
multidict==5.2.0; python_version >= '3.6'
netaddr==0.8.0
oauthlib==3.1.1; python_version >= '3.6'
openpyxl==3.0.9
pillow==8.4.0; python_version >= '3.6'
pycparser==2.21
pyjwt[crypto]==2.3.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2021.3
ratelim==0.1.6
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.1; python_version >= '3.0'
sqlparse==0.4.2; python_version >= '3.5'
tablib[xls,xlsx]==3.1.0; python_version >= '3.6'
telepath==0.2; python_version >= '3.5'
typing-extensions==4.0.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
wagtail-cache==1.0.2
wagtail-import-export==0.2
wagtail-seo==1.0.0; python_version >= '3.6'
wagtail==2.14.2; python_version >= '3.6'
wagtailfontawesome==1.2.1
webencodings==0.5.1
willow==1.4
xlrd==2.0.1
xlsxwriter==1.4.5
xlwt==1.3.0
yarl==1.7.2; python_version >= '3.6'
